added the funcitonalite of add property and experience

+++ experience-service/src/main/java/com/jabador/experience_service/repository/TimeSlotRepository.java
@@ -5,5 +5,5 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface TimeSlotRepository extends JpaRepository<TimeSlot, Double> {
+public interface TimeSlotRepository extends JpaRepository<TimeSlot, Long> {
 }
